requirements_backend.txt
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic
llama-cpp-python